Update Inheritance&Overriding
Create a class Car with fieldss make and year. Then create a child class PassengerCar  that inherits from Car and overrides the method description() to return a description of the passenger car.

Create a class Fruit with a field name. Then create a class Apple that inherits from Fruit and overrides the method taste() to return a description of the apple's taste.

Create a class Animal with fields name and age. Then create a class Dog that inherits from Animal and overrides the method sound() to return the sound a dog makes.

Create a class Item with fields name and price. Then create a class Product that inherits from Item and overrides the method description() to return a description of the product.

Create a classShape with fields color and area. Then create a class Circle that inherits from Shape and overrides the method area() to calculate the area of the circle.

Create a class Employee with fields name and salary. Then create a class Manager that inherits from Employee and overrides the method displayInfo() to display information about the manager.

Signed-off-by: PenguinSat <mukhamed.sattarov@alatoo.edu.kg>